output robocopy output to log file --> make the console far less noisy.
The result is a ~ 50% time gain: 30 seconds instead of 1mmin06 (measured on a 'hot drive', i.e. the OS has already cached the source dirtree due to previous runs shortly before; hence actual running time in the wild are expected to be slower -- nevertheless, the huge reduction in console output will always win as we've noticed with other 'noisy' applications that console stdout/stderr output is relatively costly on Win32.)

Now also includes a full copy of the tesseract train/test data as part of the snapshot.