update developer doc about NASM install -- to exactly match the behaviour expected for VSNASM.
## Footnote:

This was not done or needed on my previous dev hardware (before it went *poof!*) as that system had accrued plenty NASM cruft over the years and "just happened to work out of the box" with the original VSNASM. A new rig with a clean Win10 install + MSVC2022 Community Edition didn't fare that well, so we had to dig in and find what was ailing. This procedure works well if you do not want to clutter your box with multiple (possibly disparate) NASM installs.